Rename priv_sys_hook_exports.h to priv_sys_hooks_exports.h
